rm dup wrappers + rename
